Update triggerId query param logic.
The updated url will now always pass a triggerId through as a query
param, usually empty. Any valid Uuid will be parsed and used as a
filter; an empty string will treat the request as unfiltered, and any
other value for the triggerId param will be treated as an error on the
workflow side.